Switched to float32. Changes tests #19. Add block-attention #21 #25. Fix and improve flexibility of combiner.

+++ model/models.py
@@ -164,8 +164,8 @@ class MelTransformer(Transformer):
         if not debug:
             self.train_step = tf.function(
                 input_signature=[
-                    tf.TensorSpec(shape=(None, None, decoder_postnet.mel_channels), dtype=tf.float64),
-                    tf.TensorSpec(shape=(None, None, decoder_postnet.mel_channels), dtype=tf.float64),
+                    tf.TensorSpec(shape=(None, None, decoder_postnet.mel_channels), dtype=tf.float32),
+                    tf.TensorSpec(shape=(None, None, decoder_postnet.mel_channels), dtype=tf.float32),
                     tf.TensorSpec(shape=(None, None), dtype=tf.int64),
                     tf.TensorSpec(shape=(None), dtype=tf.float32),
                     tf.TensorSpec(shape=(None), dtype=tf.int32)
@@ -212,7 +212,7 @@ class MelTransformer(Transformer):
         # tar_inp = tar[:, :, :]
         tar_inp = tar
         # tar_real = tar[:, 1:, :]
-        tar_real = tf.concat([tar[:, 1:, :], tf.cast(tf.zeros((tf.shape(tar)[0], 1, tf.shape(tar)[-1])), tf.float64)],
+        tar_real = tf.concat([tar[:, 1:, :], tf.cast(tf.zeros((tf.shape(tar)[0], 1, tf.shape(tar)[-1])), tf.float32)],
                              axis=-2)  # shift target
         # tar_stop_prob = stop_prob[:, 1:]
         tar_stop_prob = stop_prob
@@ -287,7 +287,7 @@ class MelTextTransformer(Transformer):
         if not debug:
             self.train_step = tf.function(
                 input_signature=[
-                    tf.TensorSpec(shape=(None, None, mel_channels), dtype=tf.float64),
+                    tf.TensorSpec(shape=(None, None, mel_channels), dtype=tf.float32),
                     tf.TensorSpec(shape=(None, None), dtype=tf.int64),
                 ]
             )(self._train_step)
@@ -401,7 +401,7 @@ class TextMelTransformer(Transformer):
             self.train_step = tf.function(
                 input_signature=[
                     tf.TensorSpec(shape=(None, None), dtype=tf.int64),
-                    tf.TensorSpec(shape=(None, None, decoder_postnet.mel_channels), dtype=tf.float64),
+                    tf.TensorSpec(shape=(None, None, decoder_postnet.mel_channels), dtype=tf.float32),
                     tf.TensorSpec(shape=(None, None), dtype=tf.int64),
                     tf.TensorSpec(shape=(None), dtype=tf.float32),
                     tf.TensorSpec(shape=(None), dtype=tf.int32)
@@ -452,7 +452,7 @@ class TextMelTransformer(Transformer):
         tar_inp = tar
         # tar_real = tar[:, 1:]
         # shift target
-        tar_real = tf.concat([tar[:, 1:], tf.cast(tf.zeros((tf.shape(tar)[0], 1, tf.shape(tar)[-1])), tf.float64)], axis=-2)
+        tar_real = tf.concat([tar[:, 1:], tf.cast(tf.zeros((tf.shape(tar)[0], 1, tf.shape(tar)[-1])), tf.float32)], axis=-2)
         # tar_inp = tar[:, :-1]
         tar_stop_prob = stop_prob
         # tar_stop_prob = stop_prob[:, 1:]
